refactor(vcs/jj): extract helper functions to commands package
