+++ internal/vcs/jj/commands.go
@@ -4,9 +4,18 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
+	"net/url"
 	"os/exec"
+	"path"
+	"path/filepath"
+	"regexp"
 	"strings"
 )
+
+type trackedBookmark struct {
+	Name   string
+	Remote string
+}
 
 var ErrJJActionNotImplemented = errors.New("jj action semantics are not implemented")
 
@@ -26,6 +35,226 @@ func JJPush(path string) (string, error) {
 // be defined before enabling this operation.
 func JJPull(path string) (string, error) {
 	return "", fmt.Errorf("%w: pull has no direct Git-equivalent jj operation", ErrJJActionNotImplemented)
+}
+
+func GetRepoName(repoPath string) (string, error) {
+	return getRepoName("jj", repoPath)
+}
+
+func getRepoName(binary string, repoPath string) (string, error) {
+	remoteURL, err := getFirstRemoteURL(binary, repoPath)
+	if err != nil {
+		return "", err
+	}
+
+	if remoteURL == "" {
+		return filepath.Base(repoPath), nil
+	}
+
+	if repoName, ok := parseRepoName(remoteURL); ok {
+		return repoName, nil
+	}
+
+	return filepath.Base(repoPath), nil
+}
+
+func GetFirstRemoteURL(repoPath string) (string, error) {
+	return getFirstRemoteURL("jj", repoPath)
+}
+
+func getFirstRemoteURL(binary string, repoPath string) (string, error) {
+	output, err := runJJCommand(binary, repoPath, "git", "remote", "list")
+	if err != nil {
+		return "", err
+	}
+
+	for _, line := range strings.Split(output, "\n") {
+		line = strings.TrimSpace(line)
+		if line == "" || strings.HasPrefix(line, "Done importing changes") {
+			continue
+		}
+
+		parts := strings.Fields(line)
+		if len(parts) >= 2 {
+			return parts[1], nil
+		}
+	}
+
+	return "", nil
+}
+
+func GetBranchDisplay(repoPath string) (string, error) {
+	return getBranchDisplay("jj", repoPath)
+}
+
+func getBranchDisplay(binary string, repoPath string) (string, error) {
+	output, err := runJJCommand(
+		binary,
+		repoPath,
+		"log",
+		"-r",
+		"@",
+		"--no-graph",
+		"-T",
+		`bookmarks.join(",") ++ "|" ++ change_id.short() ++ "\n"`,
+	)
+	if err != nil {
+		return "", err
+	}
+
+	line := strings.TrimSpace(output)
+	if line == "" {
+		return "-", nil
+	}
+
+	parts := strings.SplitN(line, "|", 2)
+	if len(parts) != 2 {
+		return line, nil
+	}
+
+	if strings.TrimSpace(parts[0]) != "" {
+		return strings.TrimSpace(parts[0]), nil
+	}
+
+	if strings.TrimSpace(parts[1]) != "" {
+		return strings.TrimSpace(parts[1]), nil
+	}
+
+	return "-", nil
+}
+
+func GetUncommittedFiles(repoPath string) ([]string, error) {
+	return getUncommittedFiles("jj", repoPath)
+}
+
+func getUncommittedFiles(binary string, repoPath string) ([]string, error) {
+	output, err := runJJCommand(binary, repoPath, "diff", "--summary")
+	if err != nil {
+		return nil, err
+	}
+
+	files := []string{}
+	for _, line := range strings.Split(strings.TrimRight(output, "\n"), "\n") {
+		line = strings.TrimSpace(line)
+		if line != "" {
+			files = append(files, line)
+		}
+	}
+
+	return files, nil
+}
+
+func GetOutgoingCommits(repoPath string) ([]string, error) {
+	return getOutgoingCommits("jj", repoPath)
+}
+
+func getOutgoingCommits(binary string, repoPath string) ([]string, error) {
+	trackedBookmarks, err := getTrackedBookmarks(binary, repoPath)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(trackedBookmarks) == 0 {
+		return []string{}, nil
+	}
+
+	revset := buildTrackedOutgoingRevset(trackedBookmarks)
+	output, err := runJJCommand(
+		binary,
+		repoPath,
+		"log",
+		"-r",
+		revset,
+		"--no-graph",
+		"-T",
+		`commit_id.short() ++ "|" ++ description.first_line() ++ "\n"`,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return parseCommitSummaries(output), nil
+}
+
+func GetIncomingCommits(repoPath string) ([]string, error) {
+	return getIncomingCommits("jj", repoPath)
+}
+
+func getIncomingCommits(binary string, repoPath string) ([]string, error) {
+	trackedBookmarks, err := getTrackedBookmarks(binary, repoPath)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(trackedBookmarks) == 0 {
+		return []string{}, nil
+	}
+
+	revset := buildTrackedIncomingRevset(trackedBookmarks)
+	output, err := runJJCommand(
+		binary,
+		repoPath,
+		"log",
+		"-r",
+		revset,
+		"--no-graph",
+		"-T",
+		`commit_id.short() ++ "|" ++ description.first_line() ++ "\n"`,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	return parseCommitSummaries(output), nil
+}
+
+func getTrackedBookmarks(binary string, repoPath string) ([]trackedBookmark, error) {
+	output, err := runJJCommand(
+		binary,
+		repoPath,
+		"bookmark",
+		"list",
+		"--tracked",
+		"-T",
+		`name ++ "|" ++ remote ++ "\n"`,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	bookmarks := []trackedBookmark{}
+	seen := map[string]struct{}{}
+
+	for _, line := range strings.Split(strings.TrimRight(output, "\n"), "\n") {
+		line = strings.TrimSpace(line)
+		if line == "" {
+			continue
+		}
+
+		parts := strings.SplitN(line, "|", 2)
+		if len(parts) != 2 {
+			continue
+		}
+
+		name := strings.TrimSpace(parts[0])
+		remote := strings.TrimSpace(parts[1])
+		if name == "" || remote == "" {
+			continue
+		}
+
+		key := name + "|" + remote
+		if _, ok := seen[key]; ok {
+			continue
+		}
+		seen[key] = struct{}{}
+
+		bookmarks = append(bookmarks, trackedBookmark{
+			Name:   name,
+			Remote: remote,
+		})
+	}
+
+	return bookmarks, nil
 }
 
 // RunJJCommand executes a jj command in repoPath and returns stdout.
@@ -48,4 +277,97 @@ func runJJCommand(binary string, repoPath string, args ...string) (string, error
 	}
 
 	return stdout.String(), nil
+}
+
+func parseCommitSummaries(output string) []string {
+	commits := []string{}
+	seen := map[string]struct{}{}
+
+	for _, line := range strings.Split(strings.TrimRight(output, "\n"), "\n") {
+		line = strings.TrimSpace(line)
+		if line == "" {
+			continue
+		}
+
+		parts := strings.SplitN(line, "|", 2)
+		commitID := strings.TrimSpace(parts[0])
+		if commitID == "" {
+			continue
+		}
+		if _, ok := seen[commitID]; ok {
+			continue
+		}
+		seen[commitID] = struct{}{}
+
+		description := ""
+		if len(parts) == 2 {
+			description = strings.TrimSpace(parts[1])
+		}
+		if description == "" {
+			description = "(no description set)"
+		}
+
+		commits = append(commits, commitID+" "+description)
+	}
+
+	return commits
+}
+
+func buildTrackedOutgoingRevset(bookmarks []trackedBookmark) string {
+	parts := make([]string, 0, len(bookmarks))
+
+	for _, bookmark := range bookmarks {
+		parts = append(parts, fmt.Sprintf(
+			`(remote_bookmarks("%s", remote="%s")..bookmarks("%s"))`,
+			escapeRevsetString(bookmark.Name),
+			escapeRevsetString(bookmark.Remote),
+			escapeRevsetString(bookmark.Name),
+		))
+	}
+
+	return strings.Join(parts, " | ")
+}
+
+func buildTrackedIncomingRevset(bookmarks []trackedBookmark) string {
+	parts := make([]string, 0, len(bookmarks))
+
+	for _, bookmark := range bookmarks {
+		parts = append(parts, fmt.Sprintf(
+			`(remote_bookmarks("%s", remote="git")..remote_bookmarks("%s", remote="%s"))`,
+			escapeRevsetString(bookmark.Name),
+			escapeRevsetString(bookmark.Name),
+			escapeRevsetString(bookmark.Remote),
+		))
+	}
+
+	return strings.Join(parts, " | ")
+}
+
+func escapeRevsetString(s string) string {
+	s = strings.ReplaceAll(s, `\`, `\\`)
+	s = strings.ReplaceAll(s, `"`, `\"`)
+
+	return s
+}
+
+func parseRepoName(remote string) (string, bool) {
+	if strings.Contains(remote, ":") && strings.Contains(remote, "@") && !strings.Contains(remote, "://") {
+		parts := strings.SplitN(remote, ":", 2)
+		if len(parts) == 2 {
+			remote = "ssh://" + parts[0] + "/" + parts[1]
+		}
+	}
+
+	if u, err := url.Parse(remote); err == nil && u.Path != "" {
+		base := path.Base(u.Path)
+		base = strings.TrimSuffix(base, ".git")
+		return base, true
+	}
+
+	re := regexp.MustCompile(`([^/\\]+?)(?:\.git)?[/\\]?$`)
+	if match := re.FindStringSubmatch(remote); len(match) > 1 {
+		return match[1], true
+	}
+
+	return "", false
 }
